+++ railties/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Remove `new_framework_defaults` during `app:update` when `config.load_defaults`
+    already targets the current Rails version.
+
+    *Rune Philosof*
+
 *   Enable `config.asset_host` to read from environment by default
 
     This makes it so no code changes are needed in order to setup a CDN to